+++ lab/src/cli.js
@@ -19,7 +19,7 @@ function defaultOutDir() {
 
 function usage() {
   console.log(`
-Orbital Economy Lab v0.9.6
+Orbital Economy Lab v0.9.7
 
 Recommended workspace commands:
   lab [--input=input] [--modes=all] [--out=DIR]